orbit invariants in `src` moved to `src/orbit/invariants` directory in `@repo/sequences`
