+++ packages/sequences/src/orbit/invariants/counter.ts
@@ -1,6 +1,6 @@
 import { type Numeric } from '@repo/types/Numeric'
 
-import { type Orbit, orbit } from './orbit/orbit'
+import { type Orbit, orbit } from '../orbit'
 
 export function counter(initialState: Numeric = 0n): Orbit {
   return orbit({
