+++ packages/sequences/src/orbit/invariants/counter64.test.ts
@@ -1,7 +1,7 @@
 import { repeat } from '@repo/effects/repeat'
 
+import { type Orbit } from '../orbit'
 import { counter64 } from './counter64'
-import { type Orbit } from './orbit/orbit'
 
 describe('Counter 64 (orbit)', (): void => {
   const over64: bigint = 2n ** 64n
