+++ packages/sequences/src/orbit/invariants/counter64.ts
@@ -1,7 +1,7 @@
 import { uInt64 } from '@repo/fixed-width/bits64/uInt'
 import { type Numeric } from '@repo/types/Numeric'
 
-import { type Orbit, orbit } from './orbit/orbit'
+import { type Orbit, orbit } from '../orbit'
 
 export function counter64(initialState: Numeric = 0n): Orbit {
   return orbit({
